+++ src/e/collision.cpp
@@ -60,10 +60,10 @@ void Collision::collide(Sprite *sprite, Tilemap *map) {
       }
       
       Rect rect = {
-        .x = x * tileSize,
-        .y = y * tileSize,
-        .w = tileSize,
-        .h = tileSize
+        .x = (float) x * tileSize,
+        .y = (float) y * tileSize,
+        .w = (float) tileSize,
+        .h = (float) tileSize
       };
 
       Collision::collide(sprite, rect);
